+++ lib/dealog_backoffice/administrative_areas.ex
@@ -14,20 +14,20 @@ defmodule DealogBackoffice.AdministrativeAreas do
   def list do
     AdministrativeArea
     |> preload()
-    |> order_by([:ags])
+    |> order_by([:ars])
     |> Repo.all()
   end
 
   @doc """
-  List administrative areas by ags downwards.
+  List administrative areas by ars downwards.
   """
-  def list_hierarchical_by(ags) do
-    like_ags = "#{ags}%"
+  def list_hierarchical_by(ars) do
+    like_ars = "#{ars}%"
 
     from(a in AdministrativeArea,
-      where: a.ags == ^ags,
-      or_where: like(a.ags, ^like_ags),
-      order_by: a.ags
+      where: a.ars == ^ars,
+      or_where: like(a.ars, ^like_ars),
+      order_by: a.ars
     )
     |> preload()
     |> Repo.all()
@@ -38,8 +38,8 @@ defmodule DealogBackoffice.AdministrativeAreas do
 
   The following attributes can be passed:
 
-  - `ags` The unique AGS
-  - `parent_ags` The optional parent AGS
+  - `ars` The unique ARS
+  - `parent_ars` The optional parent ARS
   - `name` The name of the administrative area
   - `type_label` The label for the type of the administrative area
   - `type` The technical type of the administrative area (sta, lan, rbz, gem)